:recycle:Refactor: split backtest result styles file 백테스트 결과 스타일 파일 코드 스플릿
