+++ tradin/src/components/units/backtest/result/BackTestResult.styles.tsx
@@ -41,16 +41,7 @@ export const ResultInnerContainer = styled.div`
     gap: 10px;
 `;
 
-export const ResultContentContainer = styled.div`
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    width: 100%;
-    gap: 1rem;
-    height: 100%;
-    overflow-y: auto;
-    flex-shrink: 0;
-`;
+
 
 export const ResultSubtitle = styled.label`
     color: ${({ theme }) => theme.textColor};
@@ -304,113 +295,6 @@ export const CarouselDot = styled.div<{ active: boolean }>`
     cursor: pointer;
 `;
 
-// BackTest Result Content
-
-export const ResultContent = styled.div<{ strategyCount: number }>`
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 100%;
-    gap: ${props => `${0.1 / props.strategyCount}rem`};
-`;
-
-export const MetricContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 1rem;
-    height: 4rem;
-    padding: 0 1rem;
-`;
-
-export const MetricTitle = styled.h3`
-    margin: 0;
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    font-size: 0.9rem;
-    width: 7rem;
-    color: ${({ theme }) => theme.textColor};
-    z-index: 1;
-
-    @media (max-width: 799px) {
-        width: 5rem;
-        font-size: 0.8rem;
-    }
-`;
-
-export const BarChartContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    flex-grow: 1;
-    align-items: center;
-`;
-
-export const BarChartGroup = styled.div<{ strategyCount: number }>`
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-    height: ${props => `${props.strategyCount * 1.05}rem`};
-    position: relative;
-    border-radius: 0.3rem;
-    overflow: hidden;
-    gap: 0.1rem;
-`;
-
-const fillAnimation = keyframes`
-    from {
-        width: 0;
-    }
-    to {
-        width: 100%;
-    }
-`;
-
-export const BarChart = styled.div<{ width: number; color: string; isAnimating: boolean; index: number; total: number }>`
-    width: ${props => props.isAnimating ? '0' : `${props.width}%`};
-    height: ${props => `${100 / props.total}%`};
-    background-color: ${props => props.color};
-    opacity: 0.8;
-    position: absolute;
-    left: 0;
-    top: ${props => `${(props.index * 100) / props.total}%`};
-    transition: width 0.5s ease-out, opacity 0.3s ease;
-    border-radius: 0.2rem;
-    box-shadow: 0 0 0.5rem rgba(255, 255, 255, 0.1);
-    backdrop-filter: blur(3px);
-    display: flex;
-    align-items: center;
-    padding-left: 0.5rem;
-
-    ${props => props.isAnimating && css`
-        animation: ${fillAnimation} 1s ease-out forwards;
-    `}
-
-    &:hover {
-        opacity: 1;
-    }
-`;
-
-export const StrategyName = styled.span`
-    color: white;
-    font-size: 0.7rem;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    max-width: 90%;
-`;
-
-export const ResultValue = styled.div<{ strategyCount: number }>`
-    width: 6rem;
-    text-align: right;
-    color: ${({ theme }) => theme.timeTextColor};
-    font-size: ${props => `${1 - (props.strategyCount - 1) * 0.1}rem`};
-
-    @media (max-width: 799px) {
-        font-size: ${props => `${0.9 - (props.strategyCount - 1) * 0.1}rem`};
-    }
-`;
-
 // BackTest 차트 보기 방식
 
 export const ChartControls = styled.div`
@@ -439,166 +323,4 @@ export const ChartSelect = styled.select`
         font-size: 0.7rem;
         padding: 0.2rem 0.4rem;
     }
-`;
-
-// TransActionHistory
-
-export const TransactionHistoryContainer = styled.div`
-    width: 100%;
-    overflow-x: auto;
-    display: flex;
-    flex-direction: row;
-    height: 100%;
-    background-color: ${({ theme }) => theme.innerbackgroundColor};
-    padding: 0;
-    border-radius: 6px;
-    flex-shrink: 0;
-`;
-
-export const TransactionHistoryScroll = styled.div<{ strategyCount: number }>`
-    display: flex;
-    flex-direction: row;
-    width: ${props => props.strategyCount > 2 ? '150%' : '100%'};
-`;
-
-export const StrategyTransactions = styled.div<{ strategyCount: number }>`
-    display: flex;
-    flex-direction: column;
-    width: ${props => {
-        if (props.strategyCount === 1) return '100%';
-        if (props.strategyCount === 2) return '50%';
-        return '50%';
-    }};
-    padding: 0 10px;
-    height: 100%;
-`;
-export const StrategyTitle = styled.h3`
-    margin-bottom: 10px;
-    color: ${({ theme }) => theme.textColor};
-    padding: 0 1rem;
-`;
-
-export const TransactionList = styled.div<{ isSkeleton : boolean }>`
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    margin: 0.5rem;
-    background-color: ${({ theme }) => theme.moreinnerbackgroundColor};
-    border-radius: 4px;
-    padding: 1rem;
-    overflow-y: ${props => props.isSkeleton ? 'hidden' : 'auto'};
-
-    &::-webkit-scrollbar {
-        width: 8px;
-    }
-
-    &::-webkit-scrollbar-thumb {
-        background-color:  ${({ theme }) => theme.scrollbarThumbColor};
-        border-radius: 6px;
-        transition: background-color 0.3s ease;
-    }
-
-    &::-webkit-scrollbar-thumb:hover {
-        background-color: #555;
-    }
-
-    &::-webkit-scrollbar-track {
-        background: transparent;
-        border-radius: 8px;
-    }
-
-    &:hover {
-        &::-webkit-scrollbar {
-            width: 8px;
-        }
-    }
-`;
-
-export const TransactionItem = styled.div<{ strategyCount: number }>`
-    border: 1px solid ${({ theme }) => theme.borderColor};
-    border-radius: 5px;
-    padding: ${props => {
-        if (props.strategyCount === 1) return '1rem';
-        return '10px';
-    }};
-    margin-bottom: 10px;
-`;
-
-export const TransactionDetail = styled.p<{ strategyCount: number }>`
-    margin: 5px 0;
-    color: ${({ theme }) => theme.textColor};
-    font-size: ${props => {
-        if (props.strategyCount === 1) return '1em';
-        return '0.8em';
-    }};
-    font-weight: 700;
-
-    @media all and (min-width: 359px) and (max-width: 799px) {
-        font-size: ${props => {
-            if (props.strategyCount === 1) return '0.9em';
-            return '0.7em';
-        }};
-    }
-`;
-
-export const ProfitAmount = styled.span<{ isPositive: boolean }>`
-    color: ${({ isPositive, theme }) => isPositive ? theme.ProfitPositiveColor: theme.ProfitNegativeColor};
-    font-weight: bold;
-`;
-
-// BackTest 차트
-
-export const ChartContainer = styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 50vh;
-    overflow: hidden;
-    canvas {
-        width: 95% !important;
-        height: auto !important;
-    }
-    @media all and (max-width: 800px) {
-        padding: 0;
-    }
-`;
-
-// BackTest Skeleton UI
-
-export const shimmer = keyframes`
-    0% { background-position: 200% 0; }
-    100% { background-position: -200% 0; }
-`;
-
-export const SkeletonBox = styled.div`
-    background: ${({ theme }) => theme.skeletonBaseColor};
-    background-image: linear-gradient(
-        90deg,
-        ${({ theme }) => theme.skeletonBaseColor} 0%,
-        ${({ theme }) => theme.skeletonHighlightColor} 50%,
-        ${({ theme }) => theme.skeletonBaseColor} 100%
-    );
-    background-size: 200% 100%;
-    animation: ${shimmer} 2s infinite linear;
-`;
-
-export const ResultSkeletonContent = styled(SkeletonBox)`
-    width: calc(50% - 0.5rem);
-    height: 30px;
-    margin-bottom: 1rem;
-    border-radius: 4px;
-`;
-
-export const ChartSkeletonContainer = styled(SkeletonBox)`
-    width: 100%;
-    height: 50vh;
-    border-radius: 8px;
-`;
-
-export const ExecutedOptionsSkeletonBox = styled(SkeletonBox)`
-    width: 200px;
-    height: 20px;
-    border-radius: 4px;
-    align-content: center;
 `;
